+++ CHANGELOG.md
@@ -1,7 +1,7 @@
 ## Unreleased
 
 ### Added
-
+- Added create a list option to the *Random Quaternion* node
 - Added *Filter Material List By Name* node.
 - Added *Change GP Stroke Direction* node.